Add integration tests and a test case for parsing
Created an integration test to ensure that the .par files in the
samples directory compile without errors.  Doing so required
some code changes to fix bugs.

There was also an issue with the parser definition that caused
a semicolon after a whitespace terminal declaration in a
grammar file to trigger an error.  Created a test case for the
issue and updated the parser grammar to permit, but not require,
a semicolon in that location.