Massively refactor for modularity
Added more modules to attempt to work with Closure Compiler
type checking.  There is still much more work to do on that front,
and the Rhino configuration (which I have been using for testing
with Closure) is in a state of flux.  Configurations for other engines
need to catch up once that work is settled.

Added a number of tests and bug fixes as well.