Eliminate Closure Compiler warnings
Eliminated or suppressed warnings in all four of the current build
configurations.  The "global this" warning is suppressed for all four
builds.  As there is no clear way to localize this suppression, "global
this" is ignored in all locations.  The "useless code" suppression is
added for the entire Node build, as some amdclean constructs
will produce unreachable code by design.

Also fixed an issue in the default parse template in which falsy values
were conflated, which would result in incorrect parsing in certain
cases.