Clean unused files
Delete the _setup and src directories, which only had files no longer
in use.  It is possible that the webenv directory could meet a similar
fate, but more investigation is needed first.  The TODO file seemed
outdated, so it is deleted as well.  Updated bower.json to reflect
a more-current structure for future publication.