Attempt to install JDK8 during CI build
Updated continuous-integration configuration to test on both Node
4.x and Node 5.x.  Added installation of Oracle JDK8 to the
continuous-integration configuration, though whether this setting
will have the desired effect will be unknown until the next CI
build.  Removed build dependencies from the "test" Gulp target
so that the build is not repeated unnecessarily.  This change means
that "gulp test" will no longer build, so it is necessary to run one
of the other build targets first.