Ease use as package in outside builds
Updated version to 0.40.0.

Performed major refactoring to simplify use as a CommonJS-like
package in another build.  This change should help use from the
forthcoming project website, for example.

Renamed the default template from parser-driver.js.txt to
parser-driver-js.txt to work around a hard-to-find bug in requirejs
regarding passing paths to plugins.  It took several hours to
troubleshoot that one.  Bug report to requirejs/r.js is forthcoming,
time permitting.

More build components are able to be run on more platforms to
assist in troubleshooting builds in various situations.  Switched
optimizer from Rhino to Nashorn and included a comment in
gulpfile.js regarding how to debug the Nashorn process from
IntelliJ.  Now that amdclean is not necessarily run in Node --
again, to ease troubleshooting the build -- the switch to
Nashorn speeds up the build considerably over Rhino, which
is very slow to run amdclean.  Java 8 was already a requirement,
so nothing should really be lost.

The text plugin should no longer be included in optimized
build outputs, making them smaller.