Support npm versions 2 and 3
Prior builds presumed that npm packages would always be in the
same locations under the node_modules directory, which is not
the case in npm version 3.  This issue was causing builds against
Node 5 and Node 6 to fail on the CI server.

Implemented a Java-based backfill for missing JavaScript timer
globals such as setTimeout.  Not yet tested against Rhino.