Install development dependencies on npm prepublish
Attempt to fix the build by adding the --dev flag to the npm update
command in the prepublish script of package.json.  The build appears
to work on Windows without the flag, but it is failing on the
Linux-based CI server.  Add a more specific return statement to
attempt to keep a build from hanging when certain errors occur.