Use GitHub token in API call during build
Anonymous calls to the GitHub API during build sometimes fail
with HTTP 403, possibly because the rate limit for anonymous
API calls from the build-server IP address has been reached.
As a result, rhino.zip is not always downloaded, causing errors in
unit tests.  Added an encrypted environment variable with a
GitHub token (with no special permissions) to .travis.yml, and
updated gulpfile.js to detect and use the encrypted environment
variable if it is present.