Warn about GitHub API failures when building pull request
Because encrypted environment variables are not available to
pull requests on Travis CI, getting the latest Rhino version for
testing may or may not fail.  Warn instead of generating an
error if the API call fails.  The tests requiring Rhino to run will
still fail, but at least all the tests should actually be run.