Fix dependency version issues
Switch to using the google-closure-compiler npm package rather
than downloading the latest release.  Do not use the latest release
of google-closure-compiler until the additional errors it produces
have been addressed.  Do not use the latest nyc, which appears
to break the CI build for now.  Remove useless dependency on
volo, which was not working correctly on Node 6 anyway.

Update version to 0.40.2.  Update homepage in package.json.